requirements.txt
Django>=5.1.7
django-cors-headers>=4.3.1
djangorestframework>=3.14.0
psutil>=7.0.0
djangorestframework_simplejwt>=5.5.0
django-filter>=23.5